+++ tests/django_settings.py
@@ -6,6 +6,7 @@ INSTALLED_APPS = [
     "django.contrib.auth",
     "channels",
     "CypartaGraphqlSubscriptionsTools",
+    "tests.apps.TestsConfig",
 ]
 DATABASES = {"default": {"ENGINE": "django.db.backends.sqlite3", "NAME": ":memory:"}}
 DEFAULT_AUTO_FIELD = "django.db.models.BigAutoField"